- Move the deal swap status to its own component - Make sure it's mobile friendly - Resolve Doug's comments
