+++ src/dealDashboard/deal-swap-modal/deal-swap-modal.ts
@@ -1,0 +1,7 @@
+import { DealTokenSwap } from "entities/DealTokenSwap";
+import { bindable} from "aurelia-framework";
+import "./deal-swap-modal.scss";
+export class DealSwapModal {
+  @bindable deal: DealTokenSwap;
+
+}
