+++ src/dealDashboard/deal-swap-status/deal-swap-status.html
@@ -1,0 +1,46 @@
+<template>
+  <require from="../../funding/swap-status/swap-status"></require>
+  <require from="../../deals/timeLeft/timeLeft"></require>
+  <require from="../status-action-bar/status-action-bar"></require>
+  <swap-status deal.bind="deal" if.to-view="deal.primaryDao.tokens && deal.partnerDao">
+    <!-- Show this only when the deal is fully funded but before the swap has been executed -->
+    <status-action-bar
+      if.to-view="deal.isTargetReached && deal.isFunding"
+      content="IMPORTANT: Funding is complete so you must initiate swapping before the funding period expires in ${deal.timeLeftToExecute | timespan:'largest2'}."
+    >
+      <pbutton type="primary" click.delegate="executeSwap()" disabled.bind="isExecutingSwap">
+        Execute Swap <i class="fas">&#xf021;</i>
+      </pbutton>
+    </status-action-bar>
+    <!-- Show this only when the swap has been executed and the user is either the proposal lead or a representative -->
+    <status-action-bar
+      if.to-view="deal.isClaiming && !deal.isFullyClaimed && deal.isUserRepresentativeOrLead"
+      content="${deal.isRepresentativeUser ? 'Token swap has been executed and vesting is in progress.  Go To Deal Status to see the final state of the deal.' : null}"
+    >
+      <pbutton type="primary" click.delegate="router.navigate(`funding/${deal.id}`)">
+        ${deal.isRepresentativeUser ? 'Go to Claim' : 'Go to Status'}
+        <i class="fas fa-arrow-right"></i>
+      </pbutton>
+    </status-action-bar>
+    <status-action-bar
+      if.to-view="deal.isClaiming && deal.isFullyClaimed && deal.isUserRepresentativeOrLead"
+      content="All the swaps have been completed.  Go To Deal Status to see the final state of the deal"
+      icon-class="check-circle"
+      icon-color="success"
+    >
+      <pbutton type="primary" click.delegate="router.navigate(`funding/${deal.id}`)">
+        Go To Deal Status
+        <i class="fas fa-arrow-right"></i>
+      </pbutton>
+    </status-action-bar>
+    <status-action-bar
+      if.to-view="deal.isFailed && deal.isUserRepresentativeOrLead"
+      content="${deal.isRepresentativeUser ? 'IMPORTANT: You can withdraw your tokens at any time as this deal has failed.  Go To Withdraw to withdraw token deposits.' : null}"
+    >
+      <pbutton type="primary" click.delegate="router.navigate(`funding/${deal.id}`)">
+        ${deal.isRepresentativeUser ? 'Go To Withdraw' : 'Go to Status'}
+        <i class="fas fa-arrow-right"></i>
+      </pbutton>
+    </status-action-bar>
+  </swap-status>
+</template>
